refactor: migrate liquidity helpers to static class
